add timeouts for the server and requests (#2)
Using the previous context code, apply some timeouts to ensure the HTTP
server stays up under traffic.

This weaves a default 4 second per-request timeout through the handler.
This timeout covers how long it takes the HTTP handler to return data
from S3 (if it exists) or to discover the tile isn't in S3, fetch the
tile data from S3 the backend, and write it before returning. This is
configurable.

This also sets the basic HTTP server timeouts that are required for
healthy Go HTTP servers. Those aren't configurable.

A good summary of Go's HTTP server timeouts and why they're needed can
be found at https://blog.cloudflare.com/exposing-go-on-the-internet/

The `TimeoutHandler` sets the timeout on the `http.Request`'s context,
so we can use `r.Context()` in the handler to propagate the deadline to
the S3 and HTTP clients.

The values for the timeouts were chosen arbitrarily but from values that
I understand to be safe for public consumption. The
`full-request-timeout` default is the one most likely to need a change
depending on how far from S3 this server is hosted (or if the work to
write to S3 is moved to a background goroutine or whatever).

Future work: Given that we're using `http.Server` directly now, it might
be time to break out the handler func as a top-level function but I
leave that for a possible follow-up.