add tileCachingHandler to break out handler func
This moves the HTTP code out of the main func and into its own type
called `tileCachingHandler`. It's a simple refactoring that adds some
fields to the new struct to keep the pattern of adding flags in main.

I might be stepping on your toes if you're doing this in parallel and
I'm happy to take your patch instead. I'm also amenable to any changes
to the naming, docs, and layout you'd like.