mitigate thundering herd problems
This allows a single ctile process to turn multiple simultaneous
requests for the same tile it's into only a single request of each kind
needed to the backend. This request collapsing can be a pretty handy
technique for a complex caching system like ctile. We use the
singleflight library to do so.

This doesn't collapse requests across different ctile processes. That
would require more distributed systems work (using groupcache or
whatever).

The assumption here is that there will be relatively few ctile processes
running in the cluster, so the thundering herd problem being mitigated
in a single process is useful while a few hosts making singular
simultaneous requests is an okay cost.