mitigate thundering herd problems
This allows a single ctile process to turn multiple simultaneous
requests for the same tile it's into only a single request of each kind
needed to the backend. This request collapsing can be a pretty handy
technique for a complex caching system like ctile. We use the
singleflight library to do so.

This doesn't collapse requests across different ctile processes. That
would require more distributed systems work (using groupcache or
whatever). The assumption here is that there will be relatively few
ctile processes running in the cluster, so the thundering herd problem
being mitigated in a single process is useful and a few hosts making
singular simultaneous requests is currently an okay cost.

Future work to do request collapsing across ctile instances will be
easier to do with this singleflight joint in place.

We perform this request collapsing across all of the remote operations:
the initial fetch attempt from S3, the fetching from the backing CT log,
and writing to S3. Including all 3 operations at once avoids certain
race conditions. For instance, if we'd not included the initial S3
fetch, we'd hit the classic bug where two threads see that a cache entry
is empty, but one finishes filling the cache before the second has even
started the work to fill it and so the system performs the cache filling
a second time.

There is one minor caveat that every request coming into the same
singleflight.Do will have the same deadline as the request that first
requested the singleflight.Do. If something had happened to eat a lot of
time on the first request (say, garabage collection), but not the later
requests, a timeout could be returned for all of them. This is an
unlikely scenario but a possibility.

Future work could mitigate that by checking if the error returned from
the singleflight.Do was context.DeadlineExceeded while the local
goroutine's `ctx.Error` is nil and retrying if so. That felt like change
that would muddle this PR, but it feels reasonable to do later.