separate the domain logic from the HTTP work (#8)
This separates the actual fetching and caching work from its
intepretation into HTTP. This makes the thundering herd mitigations
easier to implement. The next PR will only need to modify this one new
function on `tileCachingHandler`.

Adding a `tileSource` type is unfortunate, but since the code is
currently interested in telling folks where the data came from, it's
necessary.

Similary, the `isPartialTile` handler exists because we had two
different reasons for checking the tile sizes that happened to be in one
place previously.