Make entries immutable (#18)
This fixes a race condition when concurrent requests share a
singleflight response. Previously we mutated the `contents.Entries`
field to adjust for the user's actual requested range. This is always a
race condition; sometimes it also resulted in a panic (index out of
range), when one goroutine shortened the slice and another tried to
access it.