Add passthrough handler (#16)
This passes through GET requests other than get-entries to the CT
backend. This is particularly useful for some log scanning tools that
call `/ct/v1/get-sth` before beginning their scan.

This is mainly intended as a convenience for small scale testing. In
production we plan to bypass this tool for all request paths other than
/ct/v1/get-entries.